+++ Kuplung/kuplung/settings/Settings.cpp
@@ -118,7 +118,7 @@ void Settings::saveSettings() {
     this->cfgUtils->writeString("currentFolder", this->currentFolder);
     this->cfgUtils->writeString("UIFontFile", this->UIFontFile);
 
-    this->cfgUtils->writeInt("UIFontSize", this->UIFontSize);
+    this->cfgUtils->writeInt("UIFontSize", static_cast<int>(this->UIFontSize));
     this->cfgUtils->writeInt("ModelFileParser", this->ModelFileParser);
     this->cfgUtils->writeInt("RendererType", this->RendererType);
 
@@ -148,7 +148,7 @@ void Settings::saveSettings() {
 }
 
 std::string Settings::string_format(const std::string fmt_str, ...) {
-    int final_n, n = ((int)fmt_str.size()) * 2; /* Reserve two times as much as the length of the fmt_str */
+    int final_n, n = int(fmt_str.size()) * 2; /* Reserve two times as much as the length of the fmt_str */
     std::string str;
     std::unique_ptr<char[]> formatted;
     va_list ap;
